Event refactor
Moved to separate file, tickets and purchases will be moved Soon:tm:
